Add files via upload
Hey guys, these are my updated files. I'm using a local search as a basic template in order to test the components. So far, so good. The WebApp has a search functionality that displays data I have locally stored in the program as a json file. The hard part will be merging it, having it call the back end data and hosting it on the server. I'm still adding more functionality and will be updating on github.